XWIKI-17895: Modularize the mentions data consumers
