+++ xwiki-platform-core/xwiki-platform-mentions/xwiki-platform-mentions-default/src/main/resources/META-INF/components.txt
@@ -2,6 +2,7 @@ org.xwiki.mentions.internal.async.DefaultMentionsBlockingQueueProvider
 org.xwiki.mentions.internal.async.DefaultMentionsThreadsProvider
 org.xwiki.mentions.internal.UserMentionEventListener
 org.xwiki.mentions.internal.DefaultMentionsConfiguration
+org.xwiki.mentions.internal.DefaultMentionsFormatter
 org.xwiki.mentions.internal.UserMentionsFormatter
 org.xwiki.mentions.internal.DefaultMentionsDataConsumer
 org.xwiki.mentions.internal.DefaultMentionsEventExecutor
