+++ xwiki-platform-core/xwiki-platform-mentions/xwiki-platform-mentions-default/src/test/java/org/xwiki/mentions/internal/UserMentionsFormatterTest.java
@@ -43,7 +43,7 @@ import static org.xwiki.mentions.DisplayStyle.LOGIN;
  * Test of {@link UserMentionsFormatter}.
  *
  * @version $Id$
- * @since 12.6
+ * @since 12.10RC1
  */
 @ComponentTest
 class UserMentionsFormatterTest
